# docs/openapi.json
{"openapi":"3.0.0","paths":{"/api/v1/sources/{id}/capture":{"post":{"operationId":"ManualSourceCaptureController_capture","summary":"Actualizar manualmente una fuente RSS","parameters":[{"name":"id","required":true,"in":"path","description":"Identificador estable de la fuente RSS","schema":{"type":"string"}}],"responses":{"200":{"description":"Captura finalizada; itemsParsed no representa filas persistidas","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManualSourceCaptureResponseDto"}}}},"404":{"description":"La fuente RSS no existe"},"409":{"description":"La fuente está inactiva u ocupada"},"500":{"description":"No fue posible completar la captura"},"502":{"description":"No fue posible obtener o interpretar la fuente RSS"},"504":{"description":"La captura excedió el tiempo permitido"}},"tags":["sources"]}},"/api/v1/sources/capture":{"post":{"operationId":"MultipleSourceCaptureController_capture","summary":"Actualizar manualmente múltiples fuentes RSS","description":"Captura manualmente múltiples fuentes RSS seleccionadas. Requiere al menos dos sourceIds efectivos, deduplica conservando el orden de su primera aparición y procesa cada fuente de forma aislada. Los fallos individuales aparecen en results sin convertir un batch válido en error HTTP.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MultipleSourceCaptureRequestDto"}}}},"responses":{"200":{"description":"Resultado aislado por cada fuente efectiva; los fallos individuales no cambian el 200","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MultipleSourceCaptureResponseDto"}}}},"400":{"description":"Body inválido o menos de dos sourceIds efectivos; por ejemplo [A] o [A, A]"},"500":{"description":"Fallo global inesperado fuera del aislamiento por fuente"}},"tags":["sources"]}},"/api/v1/sources":{"post":{"operationId":"SourcesController_create","summary":"Registrar una fuente RSS","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSourceDto"}}}},"responses":{"201":{"description":"Fuente creada y activa","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SourceResponseDto"}}}},"400":{"description":"URL inválida, no permitida o inaccesible"},"409":{"description":"La URL normalizada ya está registrada"}},"tags":["sources"]},"get":{"operationId":"SourcesController_list","summary":"Listar fuentes RSS","parameters":[{"name":"active","required":false,"in":"query","description":"Filtra estrictamente por estado","schema":{"enum":["true","false"],"type":"string"}}],"responses":{"200":{"description":"Listado de fuentes","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SourceResponseDto"}}}}},"400":{"description":"El filtro active no es true ni false"}},"tags":["sources"]}},"/api/v1/sources/{id}":{"get":{"operationId":"SourcesController_findById","summary":"Consultar una fuente RSS","parameters":[{"name":"id","required":true,"in":"path","description":"Identificador estable de la fuente","schema":{"type":"string"}}],"responses":{"200":{"description":"Fuente encontrada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SourceResponseDto"}}}},"404":{"description":"La fuente no existe"}},"tags":["sources"]},"put":{"operationId":"SourcesController_replace","summary":"Reemplazar la URL editable de una fuente RSS","parameters":[{"name":"id","required":true,"in":"path","description":"Identificador estable de la fuente","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplaceSourceDto"}}}},"responses":{"200":{"description":"Fuente reemplazada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SourceResponseDto"}}}},"400":{"description":"URL inválida, no permitida o inaccesible"},"404":{"description":"La fuente no existe"},"409":{"description":"La URL normalizada ya está registrada"}},"tags":["sources"]},"patch":{"operationId":"SourcesController_update","summary":"Actualizar parcialmente una fuente RSS","parameters":[{"name":"id","required":true,"in":"path","description":"Identificador estable de la fuente","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSourceDto"}}}},"responses":{"200":{"description":"Fuente actualizada o reactivada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SourceResponseDto"}}}},"400":{"description":"Cuerpo vacío o URL inválida/inaccesible"},"404":{"description":"La fuente no existe"},"409":{"description":"La URL normalizada ya está registrada"}},"tags":["sources"]},"delete":{"operationId":"SourcesController_deactivate","summary":"Desactivar una fuente RSS sin borrarla","parameters":[{"name":"id","required":true,"in":"path","description":"Identificador estable de la fuente","schema":{"type":"string"}}],"responses":{"204":{"description":"Fuente desactivada o ya inactiva"},"404":{"description":"La fuente no existe"}},"tags":["sources"]}},"/api/v1/config":{"get":{"operationId":"CaptureConfigController_getCurrent","summary":"Consultar la periodicidad global de captura","parameters":[],"responses":{"200":{"description":"Estado configurado o sin configurar","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaptureConfigResponseDto"}}}}},"tags":["config"]},"put":{"operationId":"CaptureConfigController_configure","summary":"Configurar la periodicidad global de captura","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCaptureConfigDto"}}}},"responses":{"200":{"description":"Periodicidad vigente después de la operación","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaptureConfigResponseDto"}}}},"400":{"description":"Valor ausente, inválido o fuera del catálogo"}},"tags":["config"]}},"/api/v1/news":{"get":{"operationId":"NewsController_list","parameters":[],"responses":{"200":{"description":"Listado de noticias almacenadas (lista vacía si no hay ninguna)","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NewsResponseDto"}}}}}},"tags":["news"]}},"/api/v1/dictionary":{"post":{"operationId":"DictionaryController_create","summary":"Crear una entrada del diccionario","description":"Operación administrativa sin autenticación en HU-17; debe limitarse mediante el perímetro de despliegue hasta incorporar autorización.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDictionaryEntryDto"}}}},"responses":{"201":{"description":"Entrada creada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DictionaryEntryResponseDto"}}}},"400":{"description":"Término, idioma o peso inválido"},"409":{"description":"El término canónico ya existe para el idioma"},"500":{"description":"Fallo interno controlado"}},"tags":["dictionary"]},"get":{"operationId":"DictionaryController_list","summary":"Listar el diccionario","description":"Operación administrativa sin autenticación en HU-17; debe limitarse mediante el perímetro de despliegue hasta incorporar autorización.","parameters":[],"responses":{"200":{"description":"Listado completo y ordenado","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DictionaryEntryResponseDto"}}}}},"500":{"description":"Fallo interno controlado"}},"tags":["dictionary"]}},"/api/v1/dictionary/{id}":{"get":{"operationId":"DictionaryController_findById","summary":"Consultar una entrada del diccionario","description":"Operación administrativa sin autenticación en HU-17; debe limitarse mediante el perímetro de despliegue hasta incorporar autorización.","parameters":[{"name":"id","required":true,"in":"path","schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"description":"Entrada encontrada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DictionaryEntryResponseDto"}}}},"400":{"description":"UUID mal formado"},"404":{"description":"Entrada inexistente"},"500":{"description":"Fallo interno controlado"}},"tags":["dictionary"]},"patch":{"operationId":"DictionaryController_update","summary":"Actualizar una entrada del diccionario","description":"Operación administrativa sin autenticación en HU-17; debe limitarse mediante el perímetro de despliegue hasta incorporar autorización.","parameters":[{"name":"id","required":true,"in":"path","schema":{"format":"uuid","type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDictionaryEntryDto"}}}},"responses":{"200":{"description":"Entrada actualizada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DictionaryEntryResponseDto"}}}},"400":{"description":"UUID, cuerpo, término, idioma o peso inválido"},"404":{"description":"Entrada inexistente"},"409":{"description":"El término canónico ya existe para el idioma"},"500":{"description":"Fallo interno controlado"}},"tags":["dictionary"]},"delete":{"operationId":"DictionaryController_delete","summary":"Eliminar una entrada del diccionario","description":"Operación administrativa sin autenticación en HU-17; debe limitarse mediante el perímetro de despliegue hasta incorporar autorización.","parameters":[{"name":"id","required":true,"in":"path","schema":{"format":"uuid","type":"string"}}],"responses":{"204":{"description":"Entrada eliminada físicamente"},"400":{"description":"UUID mal formado"},"404":{"description":"Entrada inexistente"},"500":{"description":"Fallo interno controlado"}},"tags":["dictionary"]}},"/api/v1/sentiment/analyze":{"post":{"operationId":"SentimentController_analyze","summary":"Analizar directamente el sentimiento de un texto","description":"Analiza un texto efímero con idioma explícito mediante el diccionario vigente; no guarda el texto ni crea noticias.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeTextRequestDto"}}}},"responses":{"200":{"description":"Análisis completado, también cuando no hay coincidencias","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SentimentAnalysisResponseDto"}}}},"400":{"description":"JSON, texto o idioma inválido"},"413":{"description":"El texto supera 10.000 puntos de código Unicode"},"500":{"description":"Fallo interno controlado"},"503":{"description":"El diccionario no está disponible"}},"tags":["sentiment"]}}},"info":{"title":"HumWorld API","description":"API REST de HumWorld","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ManualSourceCaptureResponseDto":{"type":"object","properties":{"sourceId":{"type":"string","description":"Identificador estable de la fuente capturada","example":"0f826bb6-df8e-42c5-bd57-27a213ff2f24"},"status":{"type":"string","description":"Estado de finalización de la captura","enum":["completed"],"example":"completed"},"itemsParsed":{"type":"number","description":"Cantidad de ítems interpretados del RSS; no representa filas persistidas","example":3,"minimum":0}},"required":["sourceId","status","itemsParsed"]},"CompletedMultipleSourceCaptureResultDto":{"type":"object","properties":{"sourceId":{"type":"string","description":"Identificador estable de la fuente capturada","example":"0f826bb6-df8e-42c5-bd57-27a213ff2f24"},"status":{"type":"string","enum":["completed"],"example":"completed"},"itemsParsed":{"type":"number","description":"Ítems RSS interpretados; no representa filas persistidas","example":3,"minimum":0}},"required":["sourceId","status","itemsParsed"]},"FailedMultipleSourceCaptureResultDto":{"type":"object","properties":{"sourceId":{"type":"string","description":"Identificador estable solicitado para la fuente","example":"b5f51f91-0b48-45e5-a926-2e9f89f109dc"},"status":{"type":"string","enum":["failed"],"example":"failed"},"errorCode":{"type":"string","description":"Categoría estable y sanitizada del fallo individual","enum":["source-not-found","source-inactive","source-busy","fetch/upstream","parse/invalid-rss","timeout","unexpected"],"example":"source-inactive"}},"required":["sourceId","status","errorCode"]},"MultipleSourceCaptureRequestDto":{"type":"object","properties":{"sourceIds":{"description":"Identificadores de fuentes RSS; se deduplican preservando orden y deben quedar al menos dos efectivos","example":["0f826bb6-df8e-42c5-bd57-27a213ff2f24","b5f51f91-0b48-45e5-a926-2e9f89f109dc","0f826bb6-df8e-42c5-bd57-27a213ff2f24"],"type":"array","items":{"type":"string"}}},"required":["sourceIds"]},"MultipleSourceCaptureResponseDto":{"type":"object","properties":{"results":{"type":"array","description":"Resultado por sourceId efectivo, en el orden de su primera aparición","example":[{"sourceId":"0f826bb6-df8e-42c5-bd57-27a213ff2f24","status":"completed","itemsParsed":3},{"sourceId":"b5f51f91-0b48-45e5-a926-2e9f89f109dc","status":"failed","errorCode":"source-inactive"}],"items":{"oneOf":[{"$ref":"#/components/schemas/CompletedMultipleSourceCaptureResultDto"},{"$ref":"#/components/schemas/FailedMultipleSourceCaptureResultDto"}],"discriminator":{"propertyName":"status","mapping":{"completed":"#/components/schemas/CompletedMultipleSourceCaptureResultDto","failed":"#/components/schemas/FailedMultipleSourceCaptureResultDto"}}}}},"required":["results"]},"CreateSourceDto":{"type":"object","properties":{"url":{"type":"string","description":"URL HTTP/HTTPS accesible de la fuente RSS","example":"https://example.org/feed.xml"}},"required":["url"]},"SourceResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"Identificador UUID estable de la fuente","example":"0f826bb6-df8e-42c5-bd57-27a213ff2f24"},"url":{"type":"string","description":"URL normalizada de la fuente","example":"https://example.org/feed.xml"},"active":{"type":"boolean","description":"Indica si la fuente es elegible para capturas","example":true},"createdAt":{"type":"string","description":"Fecha de creación","format":"date-time","example":"2026-09-04T12:00:00.000Z"},"updatedAt":{"type":"string","description":"Fecha de última actualización","format":"date-time","example":"2026-09-04T12:00:00.000Z"}},"required":["id","url","active","createdAt","updatedAt"]},"ReplaceSourceDto":{"type":"object","properties":{"url":{"type":"string","description":"Nueva URL HTTP/HTTPS accesible de la fuente RSS","example":"https://example.org/new-feed.xml"}},"required":["url"]},"UpdateSourceDto":{"type":"object","properties":{"url":{"type":"string","description":"Nueva URL HTTP/HTTPS accesible de la fuente RSS","example":"https://example.org/updated-feed.xml"},"active":{"type":"boolean","description":"Estado de activación de la fuente","example":true}}},"CaptureConfigResponseDto":{"type":"object","properties":{"capturePeriodicityMinutes":{"type":"number","description":"Periodicidad vigente; null indica que aún no está configurada","enum":[15,30,60,360,720,1440],"nullable":true,"example":30}},"required":["capturePeriodicityMinutes"]},"UpdateCaptureConfigDto":{"type":"object","properties":{"capturePeriodicityMinutes":{"type":"number","description":"Periodicidad global de captura expresada en minutos","enum":[15,30,60,360,720,1440],"example":30}},"required":["capturePeriodicityMinutes"]},"NewsResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"Identificador interno de la noticia almacenada"},"source":{"type":"string","description":"Referencia a la fuente RSS de origen (sourceId)"},"title":{"type":"object","description":"Título de la noticia","nullable":true},"link":{"type":"object","description":"Enlace a la noticia","nullable":true},"guid":{"type":"object","description":"GUID del ítem RSS","nullable":true},"description":{"type":"object","description":"Descripción del ítem RSS","nullable":true},"pubDate":{"type":"object","description":"Fecha de publicación reportada por el feed","nullable":true},"capturedAt":{"format":"date-time","type":"string","description":"Momento en el que HumWorld almacenó la noticia"}},"required":["id","source","capturedAt"]},"CreateDictionaryEntryDto":{"type":"object","properties":{"term":{"type":"string","description":"Término o frase; se almacena en su forma canónica","example":"muy positivo"},"language":{"type":"string","enum":["es","en"],"example":"es"},"weight":{"type":"number","minimum":-5,"maximum":5,"example":4}},"required":["term","language","weight"]},"DictionaryEntryResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"Identificador UUID estable de la entrada","format":"uuid","example":"0f826bb6-df8e-42c5-bd57-27a213ff2f24"},"term":{"type":"string","description":"Término o frase en forma canónica","example":"muy positivo"},"language":{"type":"string","enum":["es","en"],"example":"es"},"weight":{"type":"number","minimum":-5,"maximum":5,"example":4}},"required":["id","term","language","weight"]},"UpdateDictionaryEntryDto":{"type":"object","properties":{"term":{"type":"string","description":"Nuevo término o frase; se almacena en su forma canónica","example":"muy positivo"},"language":{"type":"string","enum":["es","en"],"example":"en"},"weight":{"type":"number","minimum":-5,"maximum":5,"example":0}}},"AnalyzeTextRequestDto":{"type":"object","properties":{"text":{"type":"string","description":"Texto efímero que se analizará sin persistirlo","example":"La recuperación fue excelente","maxLength":10000},"language":{"type":"string","description":"Idioma explícito del texto; no se realiza detección automática","enum":["es","en"],"example":"es"}},"required":["text","language"]},"SentimentAnalysisResponseDto":{"type":"object","properties":{"score":{"type":"number","description":"Score normalizado y redondeado a cuatro decimales","minimum":-1,"maximum":1,"example":1},"matchedTerms":{"type":"number","description":"Número de ocurrencias no solapadas encontradas","minimum":0,"example":1}},"required":["score","matchedTerms"]}}}}